[e2e] Fix issue with settings and reload
While working on migrating the host/ tests to e2e non hosted mode.
I found some issue regarding settings, where only 1 setting could be
set at a time, due to `Array.toString` doing `join(',')`.

Another issue was that when we reloaded DevTools we did not wait for
it to be in a state that we can test.

Bug: none
Change-Id: I69b860f332c708c56044c14faefb4c46c720c621
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6846619
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>