RPP: pass full set of events into handlers
This CL lays the ground work for future improvements. We noticed that
the FramesHandler has some performance overhead because we 1) take a
clone of the entire trace events array and 2) have to sort it.

The long term plan is to have the trace engine itself do the sorting
once, up front, so that we do not ever have to sort in other handlers
(there are a few handlers that sort all/a subset of the events).

This CL doesn't do that, but it does update the handler API to pass in
the set of all events into the `finalize()` function. We in a follow-up
CL will turn this into a sorted set of events, which will then allow us
to remove the explicit sort in this handler, which should make a good
performance improvement (on my M1 Pro, with a 400mb trace, this can save
about 300ms).

Bug: 436491188
Change-Id: Ie9018b09787c0444571177b71d43aa11cc00c680
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6845995
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>