RPP: remove `allTraceEntries` from RendererHandler.
The `allTraceEntries` array is relatively expensive to build in the
parsing stage because we have to add multiple events to it, and then
sort it.

But, it's only used _once_ in the actual UI, to build a pie chart for an
individual event breakdown (I think that we can get rid of this too, I
have updated the code to have a TODO and bug for a follow-up bit of work).

So, we spend a bunch of time in the handler stage to build this big
array, but we only use it to get references to events for our tests.
This CL introduces a helper to get those events lazily (and then
cached). This reduces the amount of work we have to do in the parsing
stage.

Bug: 436491188
Change-Id: I0ee4bc00e794c15868ba41f71aa4f659556d2476
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6845992
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>