RPP: limit the amount of initiators to draw
On a large trace, clicking on one event drew so many initiators when I
was zoomed out that I could barely see any of the events. It's useful to
walk back through the trace, but let's put a limit on it for both visual
and performance reasons.

Bug: 438484658
Change-Id: Ia1efd78f762bdb76852b075378a89f59955d7680
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6845964
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>