RPP: avoid costly calculation of trace entries when showing event stats
This CL updates the code that generates the pie chart statistics when an
individual event is selected.

Previously, it got a list of all trace entries and walked through that
to find child events of the selected event. However, we can do this much
more easily and with less work: we have an array of all the events on a
particular thread.

Now, we do it using the tree structure we build in parsing. We walk
through the selected node and all its children recursively using BFS to
calculate the totals as we go.

This saves us some effort, on a medium sized trace:

BEFORE: 75ms to generate pie chart data
AFTER: 5ms to generate pie chart data.

So this CL reduces the cost by ~90%.

R=alinavarkki@chromium.org

Bug: 436491188
Change-Id: Idb2a4d8f9199bffc28207aa107dd922b8ad8ea35
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6845661
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>