RPP: fix CPU throttling visual inconsistency
In crrev.com/c/6822159 the Network Throttling dropdown was migrated to
the new UI Eng Vision and also the new visual display.

It turns out that in Timeline Panel we explicitly overrode this to
remove the border around the settings, and that CL above broke that
overriding for the network selector, meaning that the CPU dropdown
looked visually different to the network one.

This CL removes the custom styling that the Timeline Panel applied,
meaning that they visually look the same. I think it's also better to
stick to the same visual appearance as the rest of DevTools rather than
customise it.

The CL also adds some minor padding just to give things a little bit
more room.

R=pfaffe@chromium.org

Bug: none
Change-Id: I7434a6b9ec502229a65e21d156b3d5ebaaae68b4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6850646
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>