RPP: improve performance of FramesHandler by taking subset of events
The FramesHandler took a copy of all trace events, sorted them, and then
passed them into TimelineFrameModel.

This was a performance hit because sorting a large trace (my test trace
of Figma has over 1m events, and it's not even particularly huge) is
expensive.

Looking through the code, I noticed that we actually only care about a
subset of the events. On my Figma test trace, about 100k of the events
are relevant to the TimelineFrameModel, which is 10% of the total.

Therefore, this CL updates the code to pick out the relevant events in
`handleEvent` and sort just that subset.

On my test trace, this has a decent performance boost:

Before:
- parse:Frames:finalize: 250ms

After:
- parse:Frames:finalize: 65ms

R=cjamcl@chromium.org

Bug: 436491188
Change-Id: I882fb1ff052fe983653228ec94bbf3041d4c46e1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6850647
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>