[stack_trace] Add StackTraceModel
We want one fragment trie per target, so we add a simple
StackTraceModel. This automatically takes care of the life-time for us.

The current plan is for StackTraceModel to be the "trie boundary". That
is callers should not be aware that fragments are stored in a trie.

R=pfaffe@chromium.org

Bug: 433162438
Change-Id: I3fbaddce5264855f740c88011d284bba53844b44
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6854804
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>