[AI] Add new Performance agent behind experiment
Adds a new conversation type for the new Performance agent, and a button
to access it in the Performance panel toolbar (behind an experiment).

This CL implements most of the planned "high-level" data as facts, and a
few of the planned agent functions. More to come.

The only impact on existing behavior should be no longer repeating the
context as a "context" ResponseType on every user message (which seemed
unintentional).

Bypass-Check-License: the presubmit check is wrong
Bug: 425270067
Change-Id: I4002688b5a202f674c05c4f30f6fc89b296e3aa8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6819863
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>