Right clicking a network log item selects it
Previously, when no network log item is selected, right click opens the
context menu but does not select the clicked network log item. This is
implemented by intercepting the mousedown event and likely intended to
not open the request details panel on right click.

However that approach is flawed: when a network log item is selected
(via left click) and the request details panel closed, right click will
open both the context menu and the request details panel.

In this new approach, we no longer intercept the mousedown event so that
the clicked network log item will be selected and correctly highlighted.
The context menu opens correctly. The event listener to open the request
details panel checks for right click, in which case it will not open nor
close the request details panel.

R=ergunsh@chromium.org

Fixed: 368510578
Change-Id: Ia0dc8f723bef2ff677c366719ccccf26f970b9eb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6875106
Commit-Queue: Yang Guo <yangguo@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>