Remove TODO about CheckLicense from PRESUBMIT.
We follow the Blink approach now, passing `.*` as RegExp for the license
header check (basically disabling it). For DevTools we still have an
ESLint rule in place that checks the license headers of source files, so
we're good for now.

Fixed: 40277703
Change-Id: Ibe138a6329de7b35b7e62c0b5444a4f33906bd80
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6885306
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>