Fix Styles panel alignment with right-pinned drawer
The updateSidebarPosition method in ElementsPanel was using the width of
the entire InspectorView to decide the layout of the Styles sidebar.
This caused incorrect calculations when the DevTools drawer was pinned
to the right, as it reduced the available width for the Elements panel itself.
This change updates the logic to use the width of the ElementsPanel's
own splitWidget.

Fixed: 430037839
Change-Id: I696dbee1c2440bc59faaa0b1113ebec6b35b4559
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6888972
Commit-Queue: Piotr Paulski <piotrpaulski@chromium.org>
Reviewed-by: Piotr Paulski <piotrpaulski@chromium.org>
Auto-Submit: Natallia Harshunova <nharshunova@chromium.org>