RPP: do not aria announce if the user triggers settings
This CL fixes a bug where if the user changed a setting (such as "Show
custom tracks") that triggered a redraw, we would re-announce any
annotations, despite the fact that they do not change and always
remained on screen.

The diff ended up being relatively large for this change; mostly because
during my testing I also found a bug where we do not clear up event
listeners on the ModificationsManager when the trace is not visible. To
fix this I moved the event handler into a distinct bound method, and
then we can remove the listener in the `#changeView` method.

This CL also removes the debouncer for the Aria Utils, I do not remember
why it was added, but in my testing today when creating a lot of
annotations and navigating around the panel, I could not recreate the
scenario where we spam the user.

R=ergunsh@chromium.org

Bug: none
Change-Id: I99212373b17a2d26c5da27021b03a35ff39a6f88
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6888592
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>