[stack_trace] Add 'walk' method to fragment Trie
The 'walk' method can start either from the root or an arbitrary node
of the trie. The callback can return `false` if the walk should not
decend into the children of the given node.

The 'StackTraceModel' will use 'walk' to find all the affected
fragments for a specific 'scriptId' to trigger re-translation.

R=pfaffe@chromium.org

Bug: 433162438
Change-Id: I9ec1b07139b5f162e71c45f60ce5cb9f9fbb9515
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6890741
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>