Fix truncation of the file extension when saving ai conversation
When exporting a chat conversation from the AI Assistance panel, the
filename is generated from the conversation's title. The `FileManager.save`
method has a server-side limit of 64 characters for the filename.

This change introduces client-side truncation for the filename to ensure
that the full path, including the 'devtools_' prefix and '.md' suffix,
does not exceed the 64-character limit.

Bug: 441706573
Change-Id: Ief7501e31f02f08a8436398bb5483cea55ca029f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6896239
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>