Open Rehydrated Devtools session in tab or window
This CL changes the current behavior for when opening an Enhanced Trace
in the Performance Panel.

Before: A new window will always open when the user opens a saved
Enhanced Trace

After: If the Devtools are in docked mode, new session will spawn on a
new tab. If the Devtools are in undocked mode, new session will spawn on
a new window.

Bug: 432043264
Change-Id: I4bc33ab482cde0a5a1f6e544b4bfd3be46fb0ee5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6897346
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Vidal Diazleal <vidorteg@microsoft.com>