[stack_trace] Add 'MissingDebugInfo' to 'Frame's
The 'DebuggerLanguagePlugin' annotates each DebuggerModel.CallFrame
in case symbol files are missing. We want to preserve this
functionality with the new StackTrace abstraction.

Once the CallStackSidebarPane is migrated, we can move the
SDK.DebuggerModel.MissingDebugFiles type to models/stack_trace, and
remove the SDK.DebuggerModel.MissingDebugInfoDetails all-together.

Note that we use an enum on purpose. We want to move the localized
UIStrings to the UI layer.

R=pfaffe@chromium.org

Bug: 433162438
Change-Id: Id7967973f9306d0054c9b56a4a448dd28de5f612
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6903182
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Simon Zünd <szuend@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>