[bindings] Implement 'translateRawFramesStep' in language plugins
The implementation is a combination of `getFunctionInfo` and
`rawLocationToUILocation` for each frame.

As the DebuggerLanguagePlugin doesn't handle outlining, all we have to
do is make sure that inlining is handled properly, and that the
missing debug info status is forwarded similar to the implementation
in `expandCallFrames`.

R=pfaffe@chromium.org

Bug: 433162438
Change-Id: I9e4a8725ab4ef510dc354bcdf01867c016f5abd4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6905383
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>