[GeminiCLI] Add `/GEMINI.md` to .gitignore
There are two ways to add a global GEMINI.md file:
* One inside `.gemini` folder.
* One in the root folder.

We're ignoring the one in the root because we want the folks
working on the devtools-frontend codebase to customize their
usage of the Gemini CLI based on the most common tasks they are
performing at that time (For example, if they are working more
on the BUILD system changes or ESLint scripts, they can have
a GEMINI.md tailored for that).

In a later stage, we can have a common `prompts/` folder
either inside `.gemini` folder or the `agents/` folder (similar to
chromium/src setup) to share common prompts & extensions.

Bug: none
Change-Id: I7a97a207b5bbb0c474cec0c75a153691a4db82ac
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6905367
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>