Link Media adorner to specific media player
Now that we return DOMNodeId together with the playerId, we can link
to a specific media player in the Media panel from the DOM element's
"media" adorner. To accomplish this, we added two maps to keep track
of the DOMNodeId <-> PlayerId relationship, and instruct the Media
panel's MainView to select and focus on the player if found.

One caveat is that the MediaModel and its underlying Media agent is not
enabled until the MainView is initialized. The current architecture
of *only* receiving data from the backend and no way to pull also makes
it difficult to start the MainView, enable the MediaModel, and pull
all the existing players. To ensure that we have players loaded before we navigate to the panel, we added a "ready" signal via a Promise. When
we receive the first batch of player data, we resolve this promise to
unblock clicking navigations.

The next step is to investigate the feasibility of refactor the
MediaModel to make it easier for pulling.

Bug: 40269197
Change-Id: I65e9ce9a49632ec9536e05425ec29ba1f10797da
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6908484
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Eric Leese <leese@chromium.org>