Fix minimum inspector width when drawer is vertical.
When the drawer is set to vertical orientation and the inspector window
or the main tab panel is schrunk too much, the main tab panel area
starts looking weird. To fix that, we need to set a larger minimum width
for the inspector and the main tab panel itself in that case.

Fixed: 439806980
Change-Id: I0c5024884068b96ebc8e371db6d98c94e4d2238c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6905294
Commit-Queue: Piotr Paulski <piotrpaulski@chromium.org>
Reviewed-by: Natallia Harshunova <nharshunova@chromium.org>
Auto-Submit: Piotr Paulski <piotrpaulski@chromium.org>