Move NetworkTimeCalculator into models/
This needs to be outside the UI layer as part of the work to enable the
NetworkRequestFormatter to not need to reach into the NetworkPanel.

There are more improvements to make here - for example, the model
depends on i18n, but I will do those in follow-ups.

R=alexrudenko@chromium.org

Bug: 442509324, 442817305
Change-Id: I56fa0f0b124121ed95473d8d03ee4f9c9e43aa33
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6911672
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>