[GdpIntegration] Add UserBadges for triggering specific badges on actions
Instead of holding static properties and using them, we're
using badge instances. Their lifecycle and activation is explicitly
handled by the UserBadges.ts.

The reason is, I didn't want to access the `name` and `title`
from `this.constructor` because of the needed type assertion
`(this.constructor as BadgeConstructor).name`.

Note:
* Needed to add `@ts-expect-error` to `Common.ObjectWrapper.ObjectWrapper`
for being able to emit events based on the `BadgeAction`. Please see the
comment.

Bug: 436201262
Change-Id: Idddfd9fedd7f43f50aef70e564430c6f59938224
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6912926
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>