[unittest] Fix out of memory error
This CL resets the TraceLoader caches after each describe that was
attached on the global scope this usally means one reset per file
but if you have describe('A'); describe('B'); on the top level of
one file it will rest 2 times.

Because we were never clearing this at one point we ended up with over 3GB on the heap.

Bug: none
Change-Id: I267faf321c0ad136f194db26bcca3ef08469464d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6917688
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>