Remove internal suffixes from most of the files
This was done via a fixer and EsLint, the rule doesn't work 100% correctly so I will not submit it.
There were some places that needed to be reverted as they were
used in WebTest and those test are hard to migrate.

Bug: none
Change-Id: I5eb031f8f5b86163c0e5b9515ab72e7a0fcf7b08
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6867869
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>
Auto-Submit: Nikolay Vitkov <nvitkov@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>