[Gardener] Deflake can-show-wasm-scopes_test
The issue highly likely happens because we're waiting for `PAUSE_INDICATOR_SELECTOR`
but not waiting for the values to update in the scope values explicitly. So, it can be that,
the debugger is paused but the scope values are not updated yet.

This CL updates the logic to not assert on the direct scope values but wait
until the scope values are our expected values. If the test is **actually** failing,
then this will be a test that timeouts.

Fixed: 442836436
Change-Id: If011a1faeb224eec6d7123b56fd8636ef7f305d6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6919130
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>