[unittest] Fix TrackAsyncOperations loop
For some reason we ended up in a loop where the `original` was actually a stubbed instance, and in that case we blocked the Main tread as we when into an infinite loop.

Instead of spending a day or two trying to fix the old code, I chose to rewrite it to keep track of original and stubs separately.
Additionally every stub now needs to be explicitly added to the AsyncTrackedFunctions.

Bug: none
Change-Id: I04b1913d4c63be603b931957d00df8d349d0afe0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6925580
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>