[GdpIntegration] Allow re-triggering a badge on re-activation
This behavior is needed for the profile creation flow via the starter badge.

It works as follows:
* User sees the starter badge with the option to "Create profile"
* User creates a profile using this flow
* After the profile creation, we want to show the starter badge again.

However, since the starter badge is already triggered, it won't
be re-triggered here.

It's safe to remove this constraint because, on reinitialization,
`UserBadges` already checks for whether the user received a given
badge or not.

For activity badges:
* If a badge is shown to user, it means it's already awarded.
So, when this badge is re-initialized, `UserBadges` sees it through
the `getAwardedBadgeNames` call.

For the starter badge, its trigger is based on 3 conditions:
1. User has a GDP profile, the setting is on: then the starter badge is awarded.
2. User has a GDP profile, the setting is off: then the starter badge
is shown as a nudge for the "Receive badges" setting.
3. User doesn't have a GDP profile: then the starter badge
is shown as a nudge to create a profile.

* For the 1st case, the mechanism works as if this is an activity badge.
* For the 2nd case, the reinitialization occurs when the user
sets the setting as `on`. After re-initialization, we still
want to continue triggering the starter badge (so that the
user can take it as an award).
* For the 3rd case, the reinitialization occurs when the user
creates a GDP profile. After re-initialization, we want to
trigger the starter badge so that they would get notified
about succesful account creation. (The case I'm fixing in this CL)

Bug: 443998148
Change-Id: I0b3f58b8ba3dfcdc9df22a86d6969c92522c9bf2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6931554
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>