[application] Fix odd UX of "Network requests" button.
The "Network requests" button in the Service Workers view of the
Application panel was implemented in a way that it'd open the Network
panel in the drawer when the Application panel was opened in the main
tab container, and would try to track when the drawer is closed to reset
the Network panel afterwards. This is a very confusing UX however with
all kinds of weird side effects, and the way it was implemented didn't
make sense.

With the CL the "Network requests" button now works just like other
actions in DevTools that open certain requests in the Network panel,
just initializing the filter to whatever necessary, and then switching
to the Network panel.

Drive-by-fix: Make sure that the filter bar is visible in the Network
panel when switching there with some active filter to make it obvious to
the user that the view is filtered to something (in case of the "Network
requests" button, the filter will be set to all service worker served
requests).

Fixed: 326229262
Change-Id: Ie97e1e4018fb59f7ca3d4d9306d92e3ea436dfc0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6935869
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>