Resolve snapshot file path in karma server
This CL fixes snapshot testings when run from a Chromium checkout.

Instead of resolving the URL to the snapshot file from the unit test,
use a new endpoint on the karma server to get the snapshot from path
relative to the source root.

The previous way (using import.meta.url) did not work in Chromium
checkouts because for some reason the snapshot files were not being
copied to the gen folder.

Also removed the "copy to gen folder" BUILD step, as it is no longer
needed.

Fixed: 444157747
Change-Id: I1aad455dc54c239f2a81616f84768cf00ee36b0c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6936941
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>