[Dialog] Allow stacking Dialogs on top of each other
This is definitely not a recommended practice for regular dialogs.
However, if we want to open a dialog in the Settings Tab which
is implemented as a Dialog, we need to allow stacking the
dialog on top of the Settings Tab.

We've tried not using the `Dialog` widget and implemented
the dialog using the native HTML `<dialog>` element, however,
due to the nature of how "Inspector View" is rendered in docked
mode, it caused problems with positioning in docked mode.

Bug: 443052776
Change-Id: Ia0f37dd3e4733dab78471e2431ab442ca4414dca
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6939235
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>