Use SVG for Google Developer Program logo
Currently, the "Google Developer Program" branding is rendered as text
using the "Google Sans" font. This font is not installed on all
developer machines, which can lead to incorrect rendering of the logo.

This change replaces the text-based logo with an SVG asset to ensure
that the branding is displayed consistently across all environments.

Bug: 444371877
Change-Id: I654a113803a1650af82f45e041aee8a2148f719f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6939358
Auto-Submit: Natallia Harshunova <nharshunova@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Natallia Harshunova <nharshunova@chromium.org>