[GdpIntegration] Refetch GDP profile details after sign-up
Previously, we didn't have a way to get notified about a successful
sign-up and the re-rendering was working correctly because of
the polling in the SettingsScreen.

However, if the user has sync enabled already, there is no polling
happening. Because of this, after a successful sign up, we weren't
updating the SyncSection with the profile information.

This CL adds an explicit callback to `GdpSignUpDialog` for
`onSuccess` and calls it when the sign-up is successfully complete.

Bug: 445623279
Change-Id: I40314f63c12f59301bf40b74ef044a6e41d6a218
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6955229
Auto-Submit: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Natallia Harshunova <nharshunova@chromium.org>
Reviewed-by: Natallia Harshunova <nharshunova@chromium.org>