Validate that events are Complete
Some events that are typically "Complete" can occasionally be "Instant"
instead. This means they won't have a duration. Our `isX` trace event
helpers mistakenly cast these events to "Complete", which vouches that
they have a `dur` field.

This resulted in a bug where `AIQueries.longestTasks` incorrectly sorted
tasks. If any task was missing a duration, the sorting method would be
ill-defined and result in not actually returning the longest tasks.

In general we don't care about tasks that are so short as to be instant,
and the same is true most of the other "Complete" events. I've updated
the trace event helpers to reflect that.

Bug: none
Change-Id: I7e0c699e61f24e3cb93e732c551cdf839ad2711d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6961429
Auto-Submit: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Paul Irish <paulirish@chromium.org>