RPP: fix logic for pairing async events to deal with reused IDs
In July 2025 this change in Perfetto:
https://chromium.googlesource.com/external/github.com/google/perfetto.git/+/aef636b27ffbf379fd722e7798030da2c5c4d699
had the side-effect of meaning that Perfetto can and will reuse IDs for
trace events if it can. This means that begin & end event pairs that do
not overlap may have the same ID.

Our logic to pair b&e events together does not account for this, and is
causing bugs like the one attached to this CL where events get dropped
if there are ID collisions.

Because the main, most noticeable place this happens, is in the
UserTimingsHandler, I have first implemented a fix just for the
performance.measure events in this handler. We also need to apply this
fix to all places where we pair events, but I've purposefully scoped
down this CL to only this handler because:

1. I think it's easier to review the logic in one specific context.

2. This is an urgent bug that is blocking folks using the e11y API, and
   I want to backmerge it into M141. I feel more comfortable doing that
   with a fix that is localised to one handler. We can then land a fix
   for M142 that applies to all handlers.

Bug: 445614711
Change-Id: I2df09d04147a3f899036fbe5204de314fd45cfba
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6965589
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>