RPP: deal with ID reuse when pairing async events
This CL is the more generic fix that first landed specifically for
performance.measure calls in crrev.com/c/6965589.

In July 2025 this change in Perfetto:
https://chromium.googlesource.com/external/github.com/google/perfetto.git/+/aef636b27ffbf379fd722e7798030da2c5c4d699
had the side-effect of meaning that Perfetto can and will reuse IDs for
trace events if it can. This means that begin & end event pairs that do
not overlap may have the same ID.

Our logic to pair `begin` and `end` events together (along with async
instant events too) does not account for this, and is causing bugs like
the one attached to this CL where events get dropped if there are ID
collisions.

The fix is to walk th events in timestamp order; collecting begin &
instant events by ID, and then grouping them when we find an end event.
If we have two groups with the same ID, we know they must not overlap,
so we know that by the time we hit the begin event of the second group,
we know the first must have been collected by now.

Bug: 445614711
Change-Id: I0b14aefe2408d9cb24394ad2fb6b93a87064669a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6969008
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>